osd: fix couples of bugs with similar error log of 52657. log errors like MOSDPGLog::encode_payload(uint64_t): Assertion `HAVE_FEATURE(features, SERVER_NAUTILUS)'
calling ProtocolV2::connect() put the object in state of START_CONNECT and a sequence of handshakes frame(including server identity frame) is pending to be exchanged before the connection is in ready state. The race condition is then sending message by the connection should not be allowed. The point is, only when the connection is ready it could be used to message exchanges.

this should fix the folowing similar issues:
https://tracker.ceph.com/issues/57845
https://tracker.ceph.com/issues/52657
https://tracker.ceph.com/issues/52694
https://tracker.ceph.com/issues/57845
https://tracker.ceph.com/issues/53685

Signed-off-by: Ben Gao <bengao168@msn.com>